+++ src/components/PlayerView/index.tsx
@@ -155,6 +155,9 @@ export function PlayerView({
       onSourceLoaded={proxy(props.onSourceLoaded)}
       onSourceUnloaded={proxy(props.onSourceUnloaded)}
       onSourceWarning={proxy(props.onSourceWarning)}
+      onAudioAdded={proxy(props.onAudioAdded)}
+      onAudioChanged={proxy(props.onAudioChanged)}
+      onAudioRemoved={proxy(props.onAudioRemoved)}
       onSubtitleAdded={proxy(props.onSubtitleAdded)}
       onSubtitleChanged={proxy(props.onSubtitleChanged)}
       onSubtitleRemoved={proxy(props.onSubtitleRemoved)}